+++ src/js/libraru_page/section_library.js
@@ -31,6 +31,7 @@ function appendMoviesToLibrary(movies) {
     'beforeend',
     movies.map(movieCardMarkup).join('')
   );
+  refs.moviesListEl.dataset.page = page;
   initRatings();
 }
 
@@ -58,33 +59,6 @@ async function initLibrary() {
     onOpenModalFilmById(movieId);
   });
   // END
-}
-
-export async function removeMovieFromLibrary(id) {
-  const movieCardToRemove = refs.moviesListEl.querySelector(
-    `li.item-movie-card[data-id="${id}"]`
-  );
-  movieCardToRemove?.remove();
-
-  const moviesLength = refs.moviesListEl.children.length;
-
-  if (moviesLength < PAGE_SIZE * page) {
-    loaderWrapper(loadMovie(moviesLength))
-      .then(movie => {
-        if (!movie) {
-          libRefs.loadMoreBtnEl.classList.add('is-hidden');
-          return;
-        }
-
-        appendMoviesToLibrary([movie]);
-      })
-      .catch(console.log);
-  }
-
-  if (moviesLength === 0) {
-    libRefs.emptyLibEl.classList.remove('is-hidden');
-    libRefs.libCatalogEl.classList.add('is-hidden');
-  }
 }
 
 function filterMoviesByGenre(evt) {
