+++ src/js/utils/initAddToLibraryBtn.js
@@ -1,6 +1,11 @@
 import { checkMovie, removeMovie, saveMovie } from '../api/libraryAPI';
-import { removeMovieFromLibrary } from '../libraru_page/section_library';
 import { onCloseModalFilm } from '../modals/modal_film';
+import loaderWrapper from '../loader';
+import { loadMovie } from '../api/libraryAPI';
+import refs from '../refs/library-refs';
+import movieCardMarkup from '../markup/movieCardMarkup';
+
+const PAGE_SIZE = 9;
 
 /**
  * Initialize button "Add to library"
@@ -47,3 +52,34 @@ const onRemoveFromLibrary = evt => {
     removeMovieFromLibrary(id);
   }
 };
+
+async function removeMovieFromLibrary(id) {
+  const movieCardToRemove = refs.moviesListEl.querySelector(
+    `li.item-movie-card[data-id="${id}"]`
+  );
+  movieCardToRemove?.remove();
+
+  const moviesLength = refs.moviesListEl.children.length;
+  const page = refs.moviesListEl.dataset.page;
+
+  if (moviesLength < PAGE_SIZE * page) {
+    loaderWrapper(loadMovie(moviesLength))
+      .then(movie => {
+        if (!movie) {
+          refs.loadMoreBtnEl.classList.add('is-hidden');
+          return;
+        }
+
+        refs.moviesListEl.insertAdjacentHTML(
+          'beforeend',
+          movieCardMarkup(movie)
+        );
+      })
+      .catch(console.log);
+  }
+
+  if (moviesLength === 0) {
+    refs.emptyLibEl.classList.remove('is-hidden');
+    refs.libCatalogEl.classList.add('is-hidden');
+  }
+}
